# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Moov Assist Web & Mobile API
  version: 0.3.0
  description: API façade sécurisée pour Chatbase, Moov Money, KYC et assistance.
servers:
  - url: https://assist.example.ga
security:
  - bearerAuth: []
paths:
  /api/auth/otp:
    post:
      security: []
      summary: Envoyer ou vérifier un OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone: { type: string, pattern: '^0[1-7][0-9]{7}$' }
                code: { type: string, pattern: '^[0-9]{6}$' }
      responses:
        '201': { description: OTP envoyé }
        '200': { description: OTP vérifié, jetons Web/mobile émis }
        '429': { $ref: '#/components/responses/RateLimited' }
  /api/auth/refresh:
    post:
      security: []
      summary: Renouveler et faire tourner les jetons
      responses:
        '200': { description: Nouvelle paire de jetons }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /api/chat:
    post:
      security: []
      summary: Converser avec l’agent Chatbase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, maxLength: 4000 }
                conversationId: { type: string }
      responses:
        '200': { description: Réponse et conversationId }
...
  /api/moov-money/transactions:
    get:
      summary: Historique du titulaire authentifié
      responses:
        '200': { description: Liste des opérations }
    post:
      summary: Envoyer de l’argent après confirmation forte
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Transfer'
      responses:
        '201': { description: Transfert accepté }
        '400': { description: Données ou preuve non concordantes }
  /api/moov-money/airtime:
    post:
      summary: Acheter du crédit après confirmation forte
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        $ref: '#/components/requestBodies/Transfer'
      responses:
        '201': { description: Recharge acceptée }
  /api/moov-money/bills:
    post:
      summary: Payer une facture après confirmation forte
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, reference, amount, confirmationToken]
              properties:
                provider: { type: string }
                reference: { type: string }
                amount: { type: number, exclusiveMinimum: 0 }
                confirmationToken: { type: string }
      responses:
        '201': { description: Paiement accepté }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
  requestBodies:
    Transfer:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [recipient, amount, confirmationToken]
            properties:
              recipient: { type: string }
              amount: { type: number, exclusiveMinimum: 0 }
              confirmationToken: { type: string }
  responses:
    Unauthorized:
      description: Session absente, expirée ou invalide
    RateLimited:
      description: Trop de requêtes